ABI Group: generate section nodes.
This is an awful, awful hack, and only here as a record of
something that works enough.

The alternative might be to generate a custom index instead? More
research is needed, which may require using pending nodes so that
the directive can collect the contents beforehand, then generate
the requisite nodes and indices afterwards...?